+++ CHANGELOG.md
@@ -17,6 +17,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Removed
 - silo-vaults: remove skim method
 
+## [1.9.0] - 2025-02-28
+### Added
+- add method `getUsersHealth` to SiloLens
+
+## [1.8.0] - 2025-02-25
+### Added
+- SiloLens redeployment
+
 ## [1.7.0] - 2025-02-26
 ### Added
 - LBTC / scBTC market on Sonic